Un-spiking, phase 0 :-)
Took a couple days off to clear the head, now developing this test-first since
I know now what I want to do, and the spike shows it can work.

The principle with the testing is that the rake task file should be a thin wrapper
around readily-testable classes, as Rake files are hard to work with (see the code
for doing BDD'd rake testing in github.com/rick/redbox, e.g.).

Also, since the main complexity appears to be in remote vs. local functionality, I'm
going to aim to have a command-buffering layer that gathers the command stream and
issues it all at once either locally or to the remote system.  This has the advantage
of being more testable, it should be ultimately faster to execute, and provides a
decent point of indirection should we want to do other fancy things.