Specs and implementation for main class.
This is the logic behind the rake tasks, the linkage with vlad/shell for remote/
local deployments.  Helper methods, tying together configuration, etc.  This doesn't
include the logic to install hooks, which we can flesh out later.

Still need the tidbits for actually running optional deploy:post_{setup,deploy}
rake tasks, as well as the implementation of the WhiskeyDisk::Config class.