Normalizing configuration data after loading.
(Currently are hard-wiring the project name, this is the next target of
feature updates/refactoring)

This is now guaranteeing that whatever comes back from WD::Config.load_data
is in the format:

{ 'project_name' => { 'environment_name' => { <config data> } }

possibly with multiple projects and environments.  This supports the degenerate (and
potentially legacy) config file format:

repository: "..."
domain: "..."

(etc.)

i.e. a simple configuration presumably used to deploy a single project in a
single environment.

But this allows us to gang up multiple configurations in a file, so we can manage multiple
environments and multiple projects all in a single config file.