Support for staleness checks
The bundle routine will ultimately be responsible for wrapping
the finalized command in a 1- or 2-repo staleness check ready
for dispatching.

This is only meaningful for deploy:now (not deploy:setup), hence
the signaling via WhiskeyDisk.enable_staleness_checks.  This is
also only relevant if ENV['check'] is turned on, so we watch that
flag as well.

The specs are in place and/or examples are named so that the logic
should be clear as to what bundle will need to do, but will need
to put the actual conditionalizing in place next.