Removing Jeweler warning from Rakefile
Basically, if you're going to be building gems, etc., on this, and
you don't have Jeweler installed, you'll see what you need when you
read the Rakefile.  And if you can't figure it out, maybe you don't
need to be building gems. :-)

This warning is annoying for people doing everything but packaging
and not having Jeweler installed.