now run SSH commands on multiple domains, serially, synchronously
We will run commands on the domains in the order the
domains are encountered in the config file.

Also, we record the status of the runs (with a simple true/false
status flag, at least for now) in WhiskeyDisk.results, which
can be polled by the rake file, etc.

Note that we do not stop on errors, but instead simply record
status and continue deploying to remote domains.