adding support for config_target in deploy.yml
This makes it possible to share config repo data for one
target (say, 'staging') with other targets.  If you have a bunch
of targets which are configured in the same way, there's no reason
to make config repo trees for every one of those environments,
rather just reference a single config repo tree.