Adding initial integration spec suite
The suite contains a set of scenarios (deploy.yml files which act as test fixtures
for describing runs of interest).  Also included with the scenarios are two bare git
repositories (one for project, another for a configuration repo).

The integration tests are enabled by setting a non-empty value in the INTEGRATION
environment variable before running the suite.

The integration suite relies upon a locally-listening git-daemon instance to act as
a git server for project.git and config.git.  It is presumed that SSH access will be
configured on the local host.  The use of /etc/hosts entries allows for fully self-
contained testing on the local system (even when offline, e.g., on an airplane).

Note:  4 specs in the integration suite are currently failing due to the 'user@domain'
"--only" bug, which will be fixed in a later commit.