check out the specified branch on setup
We use checkout -b here to create the branch for later use by deploy.
Note that some projects will be checking out their master branches
so that checkout -b might fail (in which case we fall back to a
regular git checkout).  One might think we could just check for

  branch == 'master'

But this is not sufficient, since the main branch for a repository
does not have to be named 'master'.

This should close this issue:

  https://github.com/flogic/whiskey_disk/issues/search#issue/9

  "wd setup needs to create the branch before checking out the branch"

though it is difficult for me to test at the moment due to git
version considerations.