Make #changes work on setup
By presuming that all files have changed.  We're just dropping the recorded changes
on a setup, before running hooks.  Ideally we wouldn't record the changes at all, but
not going to address that at the moment -- we'll ultimately probably fold setup into 
deploy altogether and we can deal with it then.