de-localizing remotes for integration specs
So, before we got onto the vagrant tip, for integration testing deployments
we were doing local clones and leaving the git remote set to the local
filesystem checkout.  Now we still do the local clones, but we force the
git remote on the vagrant box to point out to our git-daemon explicitly.