fix hashify_domain_filter for ruby 1.9.2
Ruby 1.8.7 was allowing me to erroneously treat a single string
(domain name) as an array and call collect on it.  Ruby 1.9.2
did away with that behavior and exposed a bug in the filter which
"hashifies" the domain settings from the config file.

Since facon (the mocking library I have been using with bacon, the
spec library I have been using) is apparently not maintained and
doesn't work with ruby 1.9.2, I have no easy way to run the test
suite to see if everything is running in 1.9.2.  I will have to
do some work to upgrade the test suite to behave under newer
rubies.