Revert facon->mocha spec changes
Facon was not runnable on 1.9.2, so switched to mocha.  Probably
would've been better to keep those changes on a topic branch until
I've dealt with other outstanding code (which was having merge
headaches with the facon->mocha changes).

Yossef fixed facon, and the upstream facon maintainer pushed his
changes to a new gem.  Now facon works on 1.9.2 so no need for the
mocha implementation (frankly I've long had some complaints about
mocha and it's JMock silliness).